split patches on proper files
